+++ src/index.ts
@@ -20,15 +20,34 @@
  * notate(obj, 'top_level.missing.value');
  * // undefined
  */
-export function notate(source: Record<string, any>, string = ""): any {
+export function notate(
+	source: Record<string, any> | Map<string, any> | Set<any>,
+	string = "",
+): any {
 	return string
-		.replace(/\[(\d+)]/g, ".$1")
+		.replace(/^\[(\d+)]/, "$1") // Array notation first cell
+		.replace(/\[(\d+)]/g, ".$1") // Array notation rest of the cells
 		.split(".")
-		.reduce(
-			(previous: any, current: string): any =>
-				typeof previous === "object" && previous !== null
-					? previous[current]
-					: previous,
-			source,
-		);
+		.reduce((previous: any, current: string): any => {
+			if (previous instanceof Map) return previous.get(current);
+			if (previous instanceof Set) return getNthItem(previous, Number(current));
+			if (previous === null) return undefined;
+			return previous?.[current];
+		}, source);
+}
+
+/**
+ * Get nth item from Set
+ */
+function getNthItem(set: Set<any>, place: number): any {
+	if (place < 0) return undefined;
+	if (place >= set.size) return undefined;
+	const iterator = set.values();
+	let index = -1;
+	let next = iterator.next();
+	while (next.done === false) {
+		if (++index === place) return next.value;
+		next = iterator.next();
+	}
+	return undefined;
 }